Feat: 카테고리 후 알람전송 부분 구현

+++ src/main/java/com/ureca/review/domain/Review.java
@@ -63,7 +63,7 @@ public class Review extends BaseEntity {
                 .build();
     }
 
-    public Review updateReviewStar(Integer newStar) {
+    public Review updateReviewStar(Double newStar) {
         return this.toBuilder()
                 .reviewStar(newStar) // 별점만 변경
                 .build();
